+++ eddist-admin/src/repository/admin_bbs_repository.rs
@@ -5,7 +5,7 @@ use eddist_core::domain::client_info::ClientInfo;
 use sqlx::{query, query_as, types::Json, FromRow, MySqlPool};
 use uuid::Uuid;
 
-use crate::{AuthedToken, Board, CreateBoardInput, Res, Thread};
+use crate::{Board, CreateBoardInput, Res, Thread};
 
 #[async_trait::async_trait]
 pub trait AdminBbsRepository: Send + Sync {
@@ -54,10 +54,6 @@ pub trait AdminBbsRepository: Send + Sync {
         body: Option<String>,
         is_abone: Option<bool>,
     ) -> anyhow::Result<Res>;
-
-    async fn get_authed_token(&self, id: Uuid) -> anyhow::Result<AuthedToken>;
-    async fn delete_authed_token(&self, id: Uuid) -> anyhow::Result<()>;
-    async fn delete_authed_token_by_origin_ip(&self, id: Uuid) -> anyhow::Result<()>;
 }
 
 #[derive(Clone)]
@@ -795,78 +791,5 @@ impl AdminBbsRepository for AdminBbsRepositoryImpl {
             client_info: res.client_info.0.into(),
             res_order: res.res_order,
         })
-    }
-
-    async fn get_authed_token(&self, id: Uuid) -> anyhow::Result<AuthedToken> {
-        let query = query_as!(
-            AuthedToken,
-            r#"
-            SELECT
-                id AS "id!: Uuid",
-                token,
-                origin_ip,
-                reduced_origin_ip,
-                writing_ua,
-                authed_ua,
-                created_at,
-                authed_at,
-                validity AS "validity!: bool",
-                last_wrote_at
-            FROM
-                authed_tokens
-            WHERE
-                id = ?
-            "#,
-            id.as_bytes().to_vec(),
-        );
-
-        let authed_token = query.fetch_one(&self.0).await?;
-
-        Ok(authed_token)
-    }
-
-    async fn delete_authed_token(&self, id: Uuid) -> anyhow::Result<()> {
-        let query = query!(
-            r#"
-            UPDATE
-                authed_tokens
-            SET
-                validity = 0
-            WHERE
-                id = ?
-        "#,
-            id.as_bytes().to_vec(),
-        );
-
-        query.execute(&self.0).await?;
-
-        Ok(())
-    }
-
-    async fn delete_authed_token_by_origin_ip(&self, id: Uuid) -> anyhow::Result<()> {
-        let query = query!(
-            r#"
-            UPDATE
-                authed_tokens
-            SET
-                validity = 0
-            WHERE
-                id IN (
-                    SELECT id FROM (
-                        SELECT
-                            id
-                        FROM
-                            authed_tokens
-                        WHERE
-                            origin_ip = ?
-                    ) tmp      
-                )
-        "#,
-            id.as_bytes().to_vec(),
-        );
-
-        query.execute(&self.0).await?;
-
-        Ok(())
     }
 }
